# MWPW-132671: Load locales from their own file