… libs/configs/adobecom.json
… libs/scripts/locales.js