Resolves: [MWPW-133060](https://jira.corp.adobe.com/browse/MWPW-133060).

👀  **Update:** `locales.js` is now `adobecom.json` . No changes in performance reported below. 

Move locales to their own file. This enables locales to be loaded by the other clouds from here, instead of using copies they have added to their own repos. This creates a single source of truth (SSOT) of geo to RFC5646 (locale, hreflang, etc) data to be maintained in Milo.

**Test URLs:**
- Before: https://main--milo--adobecom.hlx.page/?martech=off
- After: https://import-locales--milo--hparra.hlx.page/?martech=off

**Performance**

There are always performance concerns regarding any `import` so early in the Milo call stack.
In the following example, we trace an initial (no cache) load of test URL above throttled to mimic a fast 3G connection.

<img width="846" alt="Screen Shot 2023-06-20 at 17 18 19" src="https://github.com/adobecom/milo/assets/13532/3204d641-10d9-4cd0-962e-d13a6aaf6486">

KPIs:

- DOMContentLoaded fired after 1.85 seconds
- Loaded fired after 6s
- Total javascript loaded is 79.6 kb
- Total content loaded is 385 kb

The import of `locales.js` (locales) loads 1.1 kB in 590ms. The size accounts for ~1% of our javascript and even less so for overall content.

The imports `utils.js` and `locales.js` happen concurrently, with utils taking 655ms to load, so there is only a negligible increase in time due to additional bandwidth usage.

An argument may be made about DNS Lookup costs once a cloud loads this from Milo. If the cloud is already loading utils from Milo, then the DNS Lookup is already being paid. We just need to re-enable import-style concurrency. For example:

```js
  // updated example for adobecom.json
  const [
    adobecomConfig,
    { loadArea, setConfig, loadLana },
  ] = await Promise.all([
    fetch('${miloLibs}/configs/adobecom.json`).then((r) => r.json()),
    import(`${miloLibs}/utils/utils.js`),
  ]);
  setConfig({ ...adobecomConfig, ...config });
```

Performance is similar with equivalent concurrency:

<img width="846" alt="Screen Shot 2023-06-20 at 18 46 39" src="https://github.com/adobecom/milo/assets/13532/f5167892-ca52-494b-9bd1-e9105a9b8181">

There is minimal performance cost to gain a "hot" SSOT in Milo. Changes to this SSOT will immediately benefit all clouds and reduce the engineering toil of copy/pasting a configuration for each cloud each time a change is made.