@hparra 2023-06-23
@auniverseaway This doesn't block https://github.com/adobecom/milo/pull/876. Please submit your feedback on that PR as soon as you can.

> I like the idea of moving the config object out scripts from a cleanliness perspective

This is about divergence, software regressions associated with it, and the toil needed to fix them.
> fundamental premise is a bit off: "We need a source of truth for Adobe.com geos and how the map to hreflang." we don't need this, and I would probably say we don't want this either when taken into real world scenarios.

*We do need it* for actual real world scenarios affecting production right now.

> **_Every_** project goes live with different regions / locales / languages at different times.

Not relevant. This mapping is a source of truth, not a set of feature flags. It should not suggest particular support, or be tied to migration or launches. I haven't seen anything in Milo that suggests it does this or that it has been used in this manner before.

> We don't want Milo be the source of truth for this as it would require everyone to go live at the same time...

> This doesn't even get into the complexities that some projects are **language-based**, and some projects are **region-based**.

Only Express is language based. ACOM (DC, CC, FedPub), BACOM, and HelpX all use the same geo-based directory structure expressed by this version of locales.

We are giving each cloud the *option* to use a common config that lives in Milo, because it's already there, and these clouds are already hot-loading files from this repo anyway. I can always put it somewhere else, but the problem remains that the core code and blocks that use `locales` live in Milo. That's why `locales` here there to begin with, because this is where everyone is developing these features. The Milo website itself doesn't need these locales at all!

> The projects that would use Milo's locale list would be in the overwhelming minority, and that's not a good enough reason to ship this.

ACOM (DC, CC, FedPub) + BACOM + HelpX compose the majority of Adobe.com marketing. Are you saying the primary Milo customers should find somewhere else to put Milo-specific data that they share between them?