@auniverseaway 2023-07-13
@hparra I think we touched on this in our call yesterday, but I still do not want to ship this in Milo. I do not want to encourage anyone to think Milo's locales or configs are best practice and should be used. We get a lot of things wrong and projects should not be misled into thinking Milo is right. They should be able to go-live and do their own testing for their own uses... including their own Typekit decisions. More importantly, I think this sets us up for hard dependencies between milo and consumers we don't want.

To your point, yes you can say, "you don't have to use this." but I don't think that's the right approach given how many sites would actually use it. It adds complexity when there are really only two projects that need this centralized.

> Only Express is language based.

This is not correct. blog.adobe.com (the biggest Milo consumer by far) is language based... with a weird small blend of region just for the UK. stock.adobe.com/pages/artisthub is language based.

> We do need it for actual real world scenarios affecting production right now.

I disagree with this. If it was needed, sites wouldn't be live without it. Is it cumbersome for the person that has to change 4 configs? Yeah, it is. Is it better than the hell we had in 2016 where a single config could not go live until every single sub-project did a full regression on their own content? Yes, I think it is. 

We're already starting to creep into that territory now when we caused 2 different CSOs on DC because we didn't think a change affected them.